docs(tabs): format stories
